+++ src/main/java/net/meander/subtlyd/world/level/block/state/properties/BlockStatePropertiesSD.java
@@ -6,4 +6,6 @@ import net.minecraft.world.level.block.state.properties.IntegerProperty;
 public class BlockStatePropertiesSD {
     public static final IntegerProperty SNOWLOGGED_LAYERS = IntegerProperty.create("snow_layers", 0, 8);
     public static final BooleanProperty BOTTOM_SNOWLOGGED = BooleanProperty.create("bottom_snowlogged");
+    public static final IntegerProperty POTION_LEVEL = IntegerProperty.create("level", 1, 6);
+    public static final BooleanProperty IS_HEAVY_STEW = BooleanProperty.create("is_heavy_stew");
 }